+++ docs/changelog.md
@@ -6,6 +6,7 @@
 - Adds support for custom elements on dynamic views.
 - The DSL parser will now throw an error if the enterprise is set more than once (#159).
 - Adds the ability to set name/value properties on element and relationship styles.
+- Adds support for setting element style stroke widths.
 
 ## 1.20.0 (15th August 2022)